Add files to watch for browser-sync proxy
This use of browser-sync is in parallel to Docker and Webpack-hot-reloading
and is only for the benefit of viewing progress on other devices on the
network. The proxy will reload on any file change in web/src/ and is
not preferable to using the main Docker NGINX proxy port where
hot reloading will be working.